Progress on identifying what replay is currently selected
Framework has been put into place for identifying which replay slot was been initially selected

At this point in time the logic does not work as the in game logo has been altered for an event, so a different approach will be needed to find out if we are at the main menu.

Also this identification is specific to a screen resolution of 1920x1080.

Rather than hardcoding options for different resolutions, a better approach would be to allow the user to calibrate the program for their particular resolution